app/requirements.txt
appdirs==1.4.3
appnope==0.1.0
bleach==2.0.0
click==6.7
decorator==4.0.11
entrypoints==0.2.2
Flask==0.12.2
Flask-Cors==3.0.2
google-api-python-client==1.6.2
html5lib==0.999999999
httplib2==0.10.3
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
oauth2client==4.1.0
olefile==0.44
packaging==16.8
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-google-places==1.4.0
pyzmq==16.0.2
qtconsole==4.3.0
rsa==3.4.2
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
uritemplate==3.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
